Fix plugin compile

+++ loritta-plugins/loritta-birthday-2020-event/src/main/kotlin/net/perfectdreams/loritta/plugin/lorittabirthday2020/LorittaBirthday2020.kt
@@ -153,7 +153,7 @@ object LorittaBirthday2020 {
 			points = transaction(Databases.loritta) {
 				CollectedBirthday2020Points.select {
 					CollectedBirthday2020Points.user eq id
-				}.count()
+				}.count().toInt()
 			}
 		}
