+++ loritta-plugins/profile-designs/src/main/kotlin/net/perfectdreams/loritta/plugin/profiles/designs/Christmas2019ProfileCreator.kt
@@ -65,7 +65,7 @@ class Christmas2019ProfileCreator : ProfileCreator {
 		}
 
 		var xpLocal: Long? = null
-		var localPosition: Int? = null
+		var localPosition: Long? = null
 
 		if (guild != null) {
 			val localProfile = transaction(Databases.loritta) {
@@ -164,13 +164,13 @@ class Christmas2019ProfileCreator : ProfileCreator {
 		}
 	}
 
-	fun drawReputations(user: ProfileUserInfoData, graphics: Graphics, reputations: Int) {
+	fun drawReputations(user: ProfileUserInfoData, graphics: Graphics, reputations: Long) {
 		val font = graphics.font
 
 		ImageUtils.drawCenteredString(graphics, "$reputations reps", Rectangle(634, 454, 166, 52), font)
 	}
 
-	fun drawUserInfo(user: ProfileUserInfoData, userProfile: Profile, guild: Guild?, graphics: Graphics, globalPosition: Int, localPosition: Int?, xpLocal: Long?, globalEconomyPosition: Int): Int {
+	fun drawUserInfo(user: ProfileUserInfoData, userProfile: Profile, guild: Guild?, graphics: Graphics, globalPosition: Long, localPosition: Long?, xpLocal: Long?, globalEconomyPosition: Long): Int {
 		val userInfo = mutableListOf<String>()
 		userInfo.add("Global")
